+++ python/lsst/ts/criopy/m1m3ts/CoolantPumpWidget.py
@@ -20,7 +20,6 @@
 import typing
 
 import astropy.units as u
-from asyncqt import asyncSlot
 from PySide2.QtCore import Slot
 from PySide2.QtWidgets import (
     QDoubleSpinBox,
@@ -29,6 +28,7 @@ from PySide2.QtWidgets import (
     QVBoxLayout,
     QWidget,
 )
+from qasync import asyncSlot
 
 from ..gui import (
     Ampere,