+++ python/lsst/ts/criopy/m1m3ts/MixingValveWidget.py
@@ -19,7 +19,6 @@
 
 import typing
 
-from asyncqt import asyncSlot
 from PySide2.QtCore import Slot
 from PySide2.QtWidgets import (
     QDoubleSpinBox,
@@ -29,6 +28,7 @@ from PySide2.QtWidgets import (
     QVBoxLayout,
     QWidget,
 )
+from qasync import asyncSlot
 
 from ..gui import DataFormWidget, DockWindow, Percent, TimeChart, TimeChartView, Volt
 from ..salcomm import MetaSAL, command